Added a Gateway interface which connects a now-IA StorageLoader to StorageManager w/ StorageManager only depending on the interface.
I did not change anything else; All I did was move StorageLoader to another folder and make StorageManager interact with it without violating clean architecture. It seems that saving is currently disabled (as seen by the lack of uses on UpdateSave and the lack of a main method), so I will not reenable it for now.